Add HotpepperShopResponseList to select shops
