+++ src/lib/api/HotpepperApi.ts
@@ -15,7 +15,7 @@ export class HotpepperApi {
     this.apiKey = apiKeyImpl.getHotpepperApiKey() as string;
   }
 
-  private async searchShop(hotpepperApiParams: HotpepperApiForm): Promise<Object | null> {
+  public async searchShop(hotpepperApiParams: HotpepperApiForm): Promise<Object | null> {
     // execute api
     try {
       const response = await axios.get(this.HOTPEPPER_API_ENDPOINT, {
@@ -27,29 +27,6 @@ export class HotpepperApi {
       });
 
       return response.data.results.shop;
-    } catch (error) {
-      console.log(error);
-      return null;
-    }
-  }
-
-  async searchShops(hotpepperApiParamsList: Array<HotpepperApiForm>): Promise<Object | null> {
-    // execute api
-    try {
-      let responseData = Promise.all(hotpepperApiParamsList.map(param => this.searchShop(param)));
-
-        // TODO: Hotpepper APIを採用していく場合は型を用意してレスポンスを解析する
-      let selectedShops = responseData.then(responses => {
-          let nestedResponses: Array<any> = Array.from(responses);
-          let flattenedResponses = [].concat.apply([],nestedResponses);
-          console.log('Num of all response data:', flattenedResponses.length);
-          const selectedNum = 4;
-          const selectedResponses = [...Array(selectedNum)].map(() =>
-            flattenedResponses.splice(Math.floor(Math.random() * flattenedResponses.length), 1)[0]);
-          console.log('Num of selected data:', selectedResponses.length);
-          return selectedResponses;
-        });
-      return selectedShops;
     } catch (error) {
       console.log(error);
       return null;
